doesn't look like msvc supports std::optional<T&> yet, at least not for string_view
